% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{dt2gdx}
\alias{dt2gdx}
\title{dt2gdx}
\usage{
dt2gdx(gdx, dt, name, valcol, uelcols, type, field = "l")
}
\arguments{
\item{gdx}{the gdx filename.}

\item{dt}{a data.table.}

\item{name}{name of the variable.}

\item{valcol}{name of data column.}

\item{uelcols}{vector of names with dimensions.}

\item{type}{type of symbol (variable or parameter)}

\item{field}{the field if `type == 'variable'`}
}
\description{
Save a data.table to a GAMS gdx file.
}
\examples{
# round-trip:

}
