% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{raw2dt}
\alias{raw2dt}
\title{raw2dt}
\usage{
raw2dt(full_data)
}
\arguments{
\item{full_data}{a list as given by gdxrrw::rgdx.}
}
\value{
a data.table
}
\description{
Provided the raw output from gdxrrw::rgdx, create a data.table
with the correct UEL dimensions.
}
