% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{raw2gdx}
\alias{raw2gdx}
\title{raw2gdx}
\usage{
raw2gdx(gdx, var)
}
\arguments{
\item{gdx}{the gdx filename.}

\item{var}{list of properties of a gdx symbol as provided by gdxrrw::rgdx.}
}
\description{
Save to a GAMS gdx file. Works on a named list providing
domains and data as given by gdxrrw::rgdx.
This is a *workaround* to fix bugs in the implementation of gdxrrw::wgdx,
namely the problems that domains are lost when writing the output of gdxrrw:rgdx
and that for variables, a `_field` domain has always to be given.
Using this wrapper, round-tripping data between R and gdx files should be possible.
}
