% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_parameter}
\alias{read_parameter}
\title{readParameter}
\usage{
read_parameter(fname, parname)
}
\arguments{
\item{fname}{the gdx filename.}

\item{varname}{name of the variable to load.}
}
\value{
a data.table
}
\description{
Read a parameter from a gdx file.
}
\examples{
readParameter("gdxname.gdx", "parameter_name")
}
