% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readgdx.parameter}
\alias{readgdx.parameter}
\title{readgdx.parameter}
\usage{
readgdx.parameter(fname, parname)
}
\arguments{
\item{fname}{the gdx filename.}

\item{varname}{name of the variable to load.}
}
\value{
a data.table
}
\description{
Read a parameter from a gdx file.
}
\examples{
dt <- as.data.table(mtcars, keep.rownames = T)
writegdx.parameter(test_gdx, dt, test_var, valcol="wt", uelcols="rn")
new_dt <- readgdx.parameter(test_gdx, test_var)
}
