% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readgdx.variable}
\alias{readgdx.variable}
\title{readgdx.variable}
\usage{
readgdx.variable(fname, varname, field = "l")
}
\arguments{
\item{fname}{the gdx filename.}

\item{varname}{name of the variable to load.}

\item{field}{select a field or use "all".}
}
\value{
a data.table
}
\description{
Read a variable from a gdx file.
}
\examples{
dt <- as.data.table(mtcars, keep.rownames = T)
writegdx.variable(test_gdx, dt, test_var, valcol="wt", uelcols="rn")
new_dt <- readgdx.variable(test_gdx, test_var)
}
