% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readgdx}
\alias{readgdx}
\title{readgdx}
\usage{
readgdx(fname, varname, type = "variable", field = "l")
}
\arguments{
\item{fname}{the gdx filename.}

\item{varname}{name of the variable or parameter to load.}

\item{type}{string: "variable" or "parameter".}

\item{field}{if type == variable, select a field (default="l") or use "all".}
}
\value{
a data.table
}
\description{
Read a variable or parameter from a gdx file.
}
\examples{
dt <- as.data.table(mtcars, keep.rownames = T)
writegdx(test_gdx, dt, test_var, valcol="wt", uelcols="rn", type="parameter")
new_dt <- readgdx(test_gdx, test_var, type="parameter")
}
